requirements.txt
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.5
plotly==6.3.1
seaborn==0.13.2
scipy==1.10.1
scikit-learn==1.3.2
imblearn==0.0
xgboost==2.1.4
catboost==1.2.8
pymongo==4.10.1
from_root==1.3.0
evidently==0.2.8
dill==0.4.0
PyYAML==6.0.3
python-dotenv==1.0.1
neuro_mf==0.0.5
boto3==1.37.38
mypy-boto3-s3==1.40.26
botocore==1.37.38
fastapi==0.119.0
uvicorn==0.33.0
jinja2==3.1.6
python-multipart==0.0.20
-e .

# if you are getting MongoDB erron use the cmd 

# pip uninstall -y pymongo motor mongoengine djongo
# pip install -U "pymongo>=4.7" dnspython certifi
